requirements.txt
requests>=2.28.0
stix2>=3.0.0
taxii2-client>=2.3.0
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.1.0
matplotlib>=3.6.0
folium>=0.14.0
dash>=2.7.0
plotly>=5.11.0
pyyaml>=6.0
jinja2>=3.1.0
pdfkit>=1.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
apscheduler>=3.9.0
geopy>=2.3.0
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
